requirements.txt
aniso8601==8.0.0
APScheduler==3.6.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.2
PyJWT==1.4.2
pyparsing==2.4.2
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
tzlocal==2.0.0
urllib3==1.25.5
Werkzeug==0.16.0
wireless==0.3.3
